Add QR code generator for wifi ssid and password
Kudos to https://github.com/lakhanmankani/wifi_qrcode_generator for the QR code generator.
Takes the generated password, and a fixed ssid and creates a QR code, cropped a little and saved to disk.
This resulting image will be sent to the e-Paper driver(s) and the password updated on the router, somehow.