Create a selenium script to update password
Attempts to use Netgear's SOAP calls appear to only work with enabling or disabling the guest wifi unfortunately.
Only alternative is to simulate the steps to change a guest wifi password using Selenium, and chrome driver in this case.
Hopefully, a headless equivalent can be achieved from this.
Password for the router is handled with keyring library.
Will need to organise this project so that appropriate dependencies are included correctly.
This leaves only the e-Paper step where the QR code is send to it and displayed (possibly along with the password)